2 Spellings 2 Pronunciations
An attempt at improving how the English translation deals with the kanji game involving the names of the two shrines (司紀杜 / 杜紀司).

The idea here, based on the original suggestion by Lord Thantus, is to use only one way of representing the spelling of the name, and only one way of representing the pronunciation of the name, per each shrine, and to clearly establish the relation between the spellings and their corresponding pronunciations.

As part of implementing this, the translation of the horror story told by Izumi has been reworked, with "God of Demons" being renamed as "Demon of Death". All the instances of this character being mentioned have been updated to reflect this.

Finally, the tangentially related "playing around with words" lines were removed/changed, as they lacked the intended subtlety of the original Japanese lines, and felt off.